Change the entity id type from int to long.
This allows us to drop in future the free list and meta data tables.

+++ engine-tests/src/test/java/org/terasology/entitySystem/PojoEntityManagerTest.java
@@ -373,7 +373,7 @@ public class PojoEntityManagerTest {
     @Test
     public void destructionOfUnloadedEntitiesPrevented() {
         EntityRef entity = entityManager.create();
-        int id = entity.getId();
+        long id = entity.getId();
         entityManager.deactivateForStorage(entity);
         assertTrue(entity.exists());
         entity.destroy();
